+++ rusk-prover/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Default prover mode to PLONK V3 and keep explicit V2/legacy mode override.
+
 ## [1.3.0] - 2025-04-17
 
 ### Changed